# api/gen/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    Comprehensive project management API with task tracking, reporting,
    and video resources. Provides secure access control and project organization.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Little PM API
  version: 1.0.0
servers:
- description: Production API
  url: https://app.little.pm
tags:
- description: Task management within projects
  name: Tasks
paths:
  /tasks/{taskUUID}:
    delete:
      description: Permanently remove a task
      operationId: deleteTask
      parameters:
      - description: Unique task identifier
        explode: false
        in: path
        name: taskUUID
        required: true
        schema:
          $ref: '#/components/schemas/uuid'
        style: simple
      responses:
        "204":
          description: Task deleted successfully
        "403":
          content:
            application/json:
              example:
                code: 403
                message: Insufficient permissions
              schema:
                $ref: '#/components/schemas/Error'
          description: Insufficient permissions
        "404":
          content:
            application/json:
              example:
                code: 404
                message: Project not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Resource not found
        "500":
          content:
            application/json:
              example:
                code: 500
                message: Internal server error
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      security:
      - bearerAuth: []
      summary: Delete task
      tags:
      - Tasks
      x-accepts:
      - application/json
      x-tags:
      - tag: Tasks
    get:
      description: Retrieve detailed information about a specific task
      operationId: getTask
      parameters:
      - description: Unique task identifier
        explode: false
        in: path
        name: taskUUID
        required: true
        schema:
          $ref: '#/components/schemas/uuid'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: Task details retrieved
        "403":
          content:
            application/json:
              example:
                code: 403
                message: Insufficient permissions
              schema:
                $ref: '#/components/schemas/Error'
          description: Insufficient permissions
        "404":
          content:
            application/json:
              example:
                code: 404
                message: Project not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Resource not found
        "500":
          content:
            application/json:
              example:
                code: 500
                message: Internal server error
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      security:
      - bearerAuth: []
      summary: Get task details
      tags:
      - Tasks
      x-accepts:
      - application/json
      x-tags:
      - tag: Tasks
    put:
      description: Modify existing task details
      operationId: updateTask
      parameters:
      - description: Unique task identifier
        explode: false
        in: path
        name: taskUUID
        required: true
        schema:
          $ref: '#/components/schemas/uuid'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: Task updated successfully
        "400":
          content:
            application/json:
              example:
                code: 400
                message: Validation error
                details:
                - Invalid date format
                - "Missing required field: title"
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid request parameters
        "403":
          content:
            application/json:
              example:
                code: 403
                message: Insufficient permissions
              schema:
                $ref: '#/components/schemas/Error'
          description: Insufficient permissions
        "404":
          content:
            application/json:
              example:
                code: 404
                message: Project not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Resource not found
        "500":
          content:
            application/json:
              example:
                code: 500
                message: Internal server error
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      security:
      - bearerAuth: []
      summary: Update task
      tags:
      - Tasks
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Tasks
components:
  parameters:
    limit:
      description: Maximum number of items to return
      explode: true
      in: query
      name: limit
      required: false
      schema:
        default: 20
        maximum: 100
        minimum: 1
        type: integer
      style: form
    offset:
      description: Pagination offset
      explode: true
      in: query
      name: offset
      required: false
      schema:
        default: 0
        minimum: 0
        type: integer
      style: form
    projectUUID:
      description: Unique project identifier
      explode: false
      in: path
      name: projectUUID
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
      style: simple
    taskUUID:
      description: Unique task identifier
      explode: false
      in: path
      name: taskUUID
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
      style: simple
    blueprintUUID:
      description: Unique task identifier
      explode: false
      in: path
      name: blueprintUUID
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
      style: simple
    reportUUID:
      description: Unique report identifier
      explode: false
      in: path
      name: reportUUID
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
      style: simple
    videoUUID:
      description: Unique video identifier
      explode: false
      in: path
      name: videoUUID
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
      style: simple
    roleFilter:
      description: Filter by user role
      explode: true
      in: query
      name: role
      required: false
      schema:
        enum:
        - owner
        - member
        type: string
      style: form
    taskStatusFilter:
      description: Filter tasks by status
      explode: true
      in: query
      name: status
      required: false
      schema:
        enum:
        - pending
        - in_progress
        - completed
        type: string
      style: form
    blueprintType:
      description: Blueprints by its type
      explode: true
      in: query
      name: status
      required: false
      schema:
        enum:
        - toy
        type: string
      style: form
    projectStatusFilter:
      description: Filter tasks by status
      explode: true
      in: query
      name: status
      required: false
      schema:
        enum:
        - in_progress
        - completed
        type: string
      style: form
  responses:
    BadRequest:
      content:
        application/json:
          example:
            code: 400
            message: Validation error
            details:
            - Invalid date format
            - "Missing required field: title"
          schema:
            $ref: '#/components/schemas/Error'
      description: Invalid request parameters
    UnauthorizedError:
      content:
        application/json:
          example:
            code: 401
            message: Authentication required
          schema:
            $ref: '#/components/schemas/Error'
      description: Missing or invalid authentication
    ForbiddenError:
      content:
        application/json:
          example:
            code: 403
            message: Insufficient permissions
          schema:
            $ref: '#/components/schemas/Error'
      description: Insufficient permissions
    NotFoundError:
      content:
        application/json:
          example:
            code: 404
            message: Project not found
          schema:
            $ref: '#/components/schemas/Error'
      description: Resource not found
    InternalServerError:
      content:
        application/json:
          example:
            code: 500
            message: Internal server error
          schema:
            $ref: '#/components/schemas/Error'
      description: Server error
  schemas:
    uuid:
      example: 550e8400-e29b-41d4-a716-446655440000
      format: uuid
      type: string
    Error:
      example:
        code: 404
        details:
        - Invalid UUID format
        - Missing required field
        message: Resource not found
      properties:
        code:
          example: 404
          format: int32
          type: integer
        message:
          example: Resource not found
          type: string
        details:
          example:
          - Invalid UUID format
          - Missing required field
          items:
            type: string
          type: array
      type: object
    Task:
      example:
        blueprintUUID: 550e8400-e29b-41d4-a716-446655440000
        createdAt: 2000-01-23T04:56:07.000+00:00
        dueDate: 2000-01-23T04:56:07.000+00:00
        description: Research main competitors' features
        title: Competitor Analysis
        taskUUID: 550e8400-e29b-41d4-a716-446655440000
        day: MONDAY
        projectUUID: 550e8400-e29b-41d4-a716-446655440000
        status: pending
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        taskUUID:
          example: 550e8400-e29b-41d4-a716-446655440000
          format: uuid
          type: string
        projectUUID:
          example: 550e8400-e29b-41d4-a716-446655440000
          format: uuid
          type: string
        title:
          example: Competitor Analysis
          type: string
        blueprintUUID:
          example: 550e8400-e29b-41d4-a716-446655440000
          format: uuid
          type: string
        description:
          example: Research main competitors' features
          type: string
        day:
          enum:
          - MONDAY
          - TUESDAY
          - WEDNESDAY
          - THURSDAY
          - FRIDAY
          - SATURDAY
          - SUNDAY
          type: string
        status:
          enum:
          - pending
          - in_progress
          - completed
          type: string
        dueDate:
          format: date-time
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - blueprintUUID
      - projectUUID
      - status
      - taskUUID
      - title
      type: object
    TaskUpdate:
      example:
        dueDate: 2000-01-23T04:56:07.000+00:00
        description: description
        title: title
        status: pending
      properties:
        title:
          type: string
        description:
          type: string
        status:
          enum:
          - pending
          - in_progress
          - completed
          type: string
        dueDate:
          format: date-time
          type: string
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Valid JWT token required for authenticated endpoints
      scheme: bearer
      type: http